docs: move use-case test into usecases dir
