+++ packages/rest-api-client/docs/typescript.md
@@ -8,16 +8,16 @@
 
 - A set of Kintone's record fields
 - Convenient for building the type definition of the record of your app
-- For more detailed usage, see [the example code](../src/KintoneFields/exportTypes/__checks__/usecase-field.ts)
+- For more detailed usage, see [the example code](../src/KintoneFields/exportTypes/__checks__/usecases/field.ts)
 
 ## `KintoneFormFieldProperty`
 
 - A set of Kintone's form properties
 - Convenient for building the type definition of the property of your app form
-- For more detailed usage, see [the example code](../src/KintoneFields/exportTypes/__checks__/usecase-property.ts)
+- For more detailed usage, see [the example code](../src/KintoneFields/exportTypes/__checks__/usecases/property.ts)
 
 ## `KintoneFormLayout`
 
 - A set of Kintone's form layout
 - Convenient for building the type definition of the layout of your app form
-- For more detailed usage, see [the example code](../src/KintoneFields/exportTypes/__checks__/usecase-layout.ts)
+- For more detailed usage, see [the example code](../src/KintoneFields/exportTypes/__checks__/usecases/layout.ts)
