+++ packages/rest-api-client/src/KintoneFields/exportTypes/__checks__/usecases/property.ts
@@ -2,7 +2,7 @@
   When you use this package, you can import them from package root like this:
   import { KintoneRestAPIClient, KintoneFormFieldProperty } from "@kintone/rest-api-client";
 */
-import { KintoneRestAPIClient, KintoneFormFieldProperty } from "../../../";
+import { KintoneRestAPIClient, KintoneFormFieldProperty } from "../../../..";
 
 const client = new KintoneRestAPIClient({
   /* ... */
